app/src/main/java/edu/ucsd/cse110/walkstatic/runs/Run.java
package edu.ucsd.cse110.walkstatic.runs;

import java.io.Serializable;

public class Run implements Serializable, Comparable<Run>{

    private String name;
    public Run(String name){
        this.name = name;
    }

    public String getName(){
        return this.name;
    }

    @Override
    public String toString(){
        return this.name;
    }

    @Override
    public boolean equals(Object other){
        if(other == null || !(other instanceof Run)){
            return false;
        }
        Run otherRun = (Run) other;
        return this.name.equals(otherRun.name);
    }

    @Override
    public int compareTo(Run run) {
        return name.compareTo(run.name);
    }
}
